charts JS to Typescript
