--- packages/client/src/features/charts/ChartsGenreList/GenreListTitle.jsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const GenreListTitle = styled.h3`
-  font-size: 1rem;
-  color: ${props => props.theme.colors.fontColorSub};
-`;
-
-export default GenreListTitle;
