--- packages/client/src/features/charts/ChartsGenreList/GenreListWrapper.jsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export default styled.div`
-  height: 95px;
-  transition: height 400ms;
-  &:hover {
-    height: 190px;
-  }
-`;
