--- packages/client/src/features/charts/ChartsGenreList/Wrapper.jsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export default styled.div`
-  overflow: hidden;
-  margin-bottom: 20px;
-  padding: 0 15px;
-`;
