--- packages/client/src/features/charts/ChartsTracks/index.js
@@ -1,1 +1,0 @@
-export { default } from './ChartsTracks';
